Remove access to `Buffer::release()`
Cuts the `release` function from the `Buffer` public API. If a resource needs to drop their buffer early, they still have access to the `Buffer::empty()` function to replace the buffer with a dummy value.
This also fixes an attempted double-free (introduced by me :< ) caused by the `Drop` implementation not checking for an invalid pointer. The Drop implementation has also been fixed to detect if we are dropping the dummy variable.