Staging commit
Pushes a large amount of changes to synchronisation primitives and threads.

This make both (theoretically) thread safe, but breaks the threading API - this removes the ability to run functions that take arguments, but also doesn't allow you to use lambdas so you also can't capture the environment either. This needs a rewrite to allow env-capture or arguments, and potentially move the thread model to match stratosphere threads.