Checkpoint
This seems to be a good checkpoint to get some feedback. There may be some other issues like in `hid.rs` where there are function arguments in the wrong order, but I'm not going to be able to find them just by looking at everything. It'll have to be tested somehow or automated.