Kind of unhinged.
The pid placeholder value is sometimes not there in the request even for cmif, and the aruid apparently isn't guaranteed to be the same as the PID. Therefore I have had to split the AppletResourceUserId send/receive out so it isn't sent as a ProcessId. the ProcessId is now just a sentinel value that marks the request for PID insertion by the kernel. There is now a new PidPlaceholder type that will be 8-bytes for cmif and 0 for tipc, to otherwise emulate the pid placeholder functionality.