Stage the current state
This is all currently building at running (as far as I can tell from testing with the nx examples), but I am continuing to work on it as I am trying to port a copy of libtesla. The biggest outstanding item at the moment is needing ergonomic support for `appletAE` and `appletOE` that is automatically selected via flag in `libnx` - but needs something on our side so we can do things like service detection in sysmodules.