Small changes plus docs cleanup for PR.
This adds scaling mode setting to CanvasManagers to control scaling modes when the layer and the framebuffer aren't the same size.

Adds a function to pm required to port the emuiibo overlay to rust.

Reverts the docs folder to aarch64-switch-rs/nx:master so the PR review screen in GitHub doesn't crash.