Fix underconstrained type on `nx::sf::Buffer`.
The buffer type previously didn't take into account the lifetime of the referenced data, so it is easy to cause use after free in API clients or accidentally hold on to buffer references too long in server implementations.

This gives a lifetime bound to the buffer type so it (hopefully) doesn't outlive the source data or escape the server function.