Console poc (#36)
* PoC Console for logging.

* Fix underconstrained type on `nx::sf::Buffer`.

The buffer type previously didn't take into account the lifetime of the referenced data, so it is easy to cause use after free in API clients or accidentally hold on to buffer references too long in server implementations.

This gives a lifetime bound to the buffer type so it (hopefully) doesn't outlive the source data or escape the server function.

* WIP - streaming console output is nicer, but no vty support yet.

* WIP2

* Interactive ANSI console MVP

* fix warnings.

* remove panic from raw frame writes.

* Main no longer returns a Result that automatically panics.

* Fix typo.